Force the use of full module paths and simplify utils
