+++ sysdescrparser/__init__.py
@@ -2,10 +2,4 @@
 
 """Sysdescrparser."""
 
-import sys
-
-# pylint: disable=unused-import
-if sys.version_info.major == 2:
-    from sysdescrparser import sysdescrparser
-else:
-    from sysdescrparser.sysdescrparser import sysdescrparser
+# Removed lifting to force full path use
