+++ sysdescrparser/a10.py
@@ -2,7 +2,7 @@
 
 """sysdescrparser.a10."""
 
-from sysdescr import SysDescr
+from sysdescrparser.sysdescr import SysDescr
 
 
 # pylint: disable=no-name-in-module
