+++ sysdescrparser/cisco_iosxr.py
@@ -4,7 +4,7 @@
 
 
 import re
-from cisco import Cisco
+from sysdescrparser.cisco import Cisco
 
 
 # pylint: disable=no-member
