+++ sysdescrparser/freebsd.py
@@ -3,7 +3,7 @@
 """sysdescrparser.freebsd."""
 
 import re
-from sysdescr import SysDescr
+from sysdescrparser.sysdescr import SysDescr
 
 
 # pylint: disable=no-name-in-module
