+++ sysdescrparser/paloalto.py
@@ -2,7 +2,7 @@
 
 """sysdescrparser.paloalto."""
 
-from sysdescr import SysDescr
+from sysdescrparser.sysdescr import SysDescr
 
 
 # pylint: disable=no-name-in-module
